Refactor map arrays to maps.js module
