+++ html/pacman.html
@@ -7,9 +7,7 @@
     <title>Portland Pacman Game</title>
 
     <link rel="stylesheet" type="text/css" href="../css/pacman.css" />
-    <script defer src="../javascript/pacman.js"></script>
-    <!-- <style></style> -->
-    <!-- <link rel="stylesheet" href="../styles.css" /> -->
+    <script defer type="module" src="../javascript/pacman.js"></script>
   </head>
   <body>
     <h1>Hello World</h1>
